move app/api/contests/get_administered_list.* to app/api/items
