+++ app/api/items/get_administered_list.go
@@ -1,4 +1,4 @@
-package contests
+package items
 
 import (
 	"net/http"
